Fix counter reporting for append mode
The patch to add per file counters accidentally broke the counter
reporting for append mode. This patch changes append mode to no
longer actually append the text data of the individual files but
instead introduces global state which is reset for each file for
non-append mode but is reused for append mode having the same effect
as concatenating individual files and passing the concatenated file
to pgspot.
This also fixes linenumber reporting for append mode which will
now show line numbers for the individual files.